Solution to First part of U 1
1) U1_1.py - code to display some high freq words along with % of
occurrence in our sample text (MobyDick)

2) comparative_percentage.txt - text file with a table of most frequent
words (above 4 letters) with % of occurrence in a)Moby Dick b)Literature
btw period 1800-1900.
Motivation for b) - the content of our text is characterized by
unusually frequently occurring words. Hence we compare a) and b) to find
these
Note: a) and b) are not labeled in the column headings in the txt file